move to custom task interfaces within datarobot_drum
