+++ custom_model_runner/datarobot_drum/custom_task_interfaces/regression_interface.py
@@ -1,4 +1,4 @@
-from .estimator import EstimatorInterface
+from .estimator_interface import EstimatorInterface
 
 
 class RegressionInterface(EstimatorInterface):
